Added icons/splash screens for all iOS platforms, changed project to build font in natively.
The title says it all. Android team will need to change their project
to include the font from www/css natively as well. Just drop it in
assets/fonts and you'll be good. Also, Android, feel free to life our
splash PNG files to modify for your app as well as the icon. On
Android, though, the oblong ellipse alone may look better as an icon,
without the background.